requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
dnspython==2.6.1
filelock==3.16.1
idna==3.10
requests==2.32.3
requests-file==2.1.0
tldextract==5.1.2
urllib3==2.2.3
